Modul für Koordintator anzeigen

+++ SopraMMS/src/management/ModuleAdministration.java
@@ -104,7 +104,7 @@ public class ModuleAdministration {
 		moduleDBController.createModuleByModuleManager(module);
 	}
 	
-	public Module getModuleByID(int moduleID){
+	public Module getModuleByID(long moduleID){
 		return moduleDBController.getModule(moduleID);
 	}
